Fixed bugs

+++ tracker/src/main/java/ir/intoo/api/tracker/model/Configure.kt
@@ -3,5 +3,5 @@ package ir.intoo.api.tracker.model
 class Configure {
     var accessToken: String = ""
     var runTimeIntervalSeconds: Long = 0
-    var changeLocationDetectionMeters: Long = 0
+    var changeLocationDetectionMeters: Float = 0.0F
 }
